ukernels: pack: move the generation of i64 padding_value to codegen (#13264)
#13233 changed the `pack` ukernel to take the `padding_value` as a i64.
Something has to convert from the `tensor.pack` `padding_value` (in the
element type) to i64. #13233 let that be done by the vmvx/module.c entry
point. In order to have codegen call ukernels without having to
duplicate VMVX logic, the present PR moves this to the `LowerToUKernels`
pass.

+++ compiler/src/iree/compiler/Dialect/VMVX/vmvx.imports.mlir
@@ -472,7 +472,7 @@ vm.import private @pack.f32f32(
   %out_size1 : i64,
   %out_size2 : i64,
   %out_size3 : i64,
-  %padding_value : f32,
+  %padding_value : i64,
   %flags : i32
 )
 
@@ -489,7 +489,7 @@ vm.import private @pack.i8i8(
   %out_size1 : i64,
   %out_size2 : i64,
   %out_size3 : i64,
-  %padding_value : i32,
+  %padding_value : i64,
   %flags : i32
 )
 
@@ -506,7 +506,7 @@ vm.import private @pack.i32i32(
   %in_size3 : i64,
   %out_size0 : i64,
   %out_size1 : i64,
-  %padding_value : i32,
+  %padding_value : i64,
   %flags : i32
 )
